+++ src/main/java/com/yanny/ytech/ModBusSubscriber.java
@@ -3,6 +3,7 @@ package com.yanny.ytech;
 import com.yanny.ytech.compatibility.CuriosCapability;
 import com.yanny.ytech.compatibility.TopCompatibility;
 import com.yanny.ytech.configuration.SpearType;
+import com.yanny.ytech.configuration.Utils;
 import com.yanny.ytech.configuration.block_entity.WellPulleyBlockEntity;
 import com.yanny.ytech.configuration.entity.*;
 import com.yanny.ytech.configuration.item.BasketItem;
@@ -13,6 +14,7 @@ import com.yanny.ytech.configuration.renderer.*;
 import com.yanny.ytech.configuration.screen.AqueductFertilizerScreen;
 import com.yanny.ytech.configuration.screen.PrimitiveAlloySmelterScreen;
 import com.yanny.ytech.configuration.screen.PrimitiveSmelterScreen;
+import com.yanny.ytech.configuration.screen.StrainerScreen;
 import com.yanny.ytech.network.irrigation.IrrigationServerNetwork;
 import com.yanny.ytech.network.irrigation.IrrigationUtils;
 import com.yanny.ytech.registration.YTechBlockEntityTypes;
@@ -82,6 +84,7 @@ public class ModBusSubscriber {
             MenuScreens.register(YTechMenuTypes.AQUEDUCT_FERTILIZER.get(), AqueductFertilizerScreen::new);
             MenuScreens.register(YTechMenuTypes.PRIMITIVE_ALLOY_SMELTER.get(), PrimitiveAlloySmelterScreen::new);
             MenuScreens.register(YTechMenuTypes.PRIMITIVE_SMELTER.get(), PrimitiveSmelterScreen::new);
+            MenuScreens.register(YTechMenuTypes.STRAINER.get(), StrainerScreen::new);
         });
     }
 
@@ -95,6 +98,7 @@ public class ModBusSubscriber {
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.FIRE_PIT.get(), FirePitRenderer::new);
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.MILLSTONE.get(), MillstoneRenderer::new);
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.POTTERS_WHEEL.get(), PottersWheelRenderer::new);
+        event.registerBlockEntityRenderer(YTechBlockEntityTypes.STRAINER.get(), StrainerRenderer::new);
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.TANNING_RACK.get(), TanningRackRenderer::new);
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.TOOL_RACK.get(), ToolRackRenderer::new);
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.TREE_STUMP.get(), TreeStumpRenderer::new);
@@ -143,6 +147,8 @@ public class ModBusSubscriber {
         for (SpearType spearType : SpearType.values()) {
             event.register(MODEL_IN_HAND_LOCATIONS.get(spearType));
         }
+
+        event.register(Utils.modBlockLoc("strainer_net"));
     }
 
     @SubscribeEvent