+++ src/main/java/com/yanny/ytech/configuration/block_entity/CraftingWorkspaceBlockEntity.java
@@ -106,7 +106,7 @@ public class CraftingWorkspaceBlockEntity extends BlockEntity {
                 itemList.set(index, ItemStack.EMPTY);
                 bitmask &= ~(1 << index);
                 wasChange = true;
-            } else if (itemStack.is(YTechItemTags.SHARP_FLINTS)) {
+            } else if (itemStack.is(YTechItemTags.TOOL_FOR_CRAFTING_WORKBENCH)) {
                 return constructBlock(itemStack, pLevel, pPos, pPlayer, pHand, pHit);
             }
 
@@ -115,7 +115,7 @@ public class CraftingWorkspaceBlockEntity extends BlockEntity {
                 pLevel.blockEntityChanged(pPos);
                 return InteractionResult.CONSUME;
             }
-        } else if (itemStack.is(YTechItemTags.SHARP_FLINTS)) {
+        } else if (itemStack.is(YTechItemTags.TOOL_FOR_CRAFTING_WORKBENCH)) {
             return constructBlock(itemStack, pLevel, pPos, pPlayer, pHand, pHit);
         }