+++ src/main/java/com/yanny/ytech/configuration/block_entity/MachineBlockEntity.java
@@ -42,13 +42,9 @@ public abstract class MachineBlockEntity extends BlockEntity implements MenuProv
         return containerData.getCount();
     }
 
-    public void tickServer(@NotNull Level level, @NotNull BlockPos pos, @NotNull BlockState blockState, @NotNull MachineBlockEntity blockEntity) {
+    public void tickServer(@NotNull Level level, @NotNull BlockPos pos, @NotNull BlockState blockState, @NotNull MachineBlockEntity blockEntity) {}
 
-    }
-
-    public void tickClient(@NotNull Level level, @NotNull BlockPos pos, @NotNull BlockState blockState, @NotNull MachineBlockEntity blockEntity) {
-
-    }
+    public void tickClient(@NotNull Level level, @NotNull BlockPos pos, @NotNull BlockState blockState, @NotNull MachineBlockEntity blockEntity) {}
 
     @Override
     protected void saveAdditional(@NotNull CompoundTag tag) {