+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # YTech Changelog
 
+## []
+
+- Added Strainer for obtaining Galena/Cassiterite
+- Removed generation of Galena/Cassiterite ores in world
+- Fixed missing item dropping on block destroy
+- Added custom tag for Crafting Workbench Tool
+
 ## [0.9.2]
 
 - Changed Stone Mortar and Pestle recipe